remove pointless cast to int
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@22464 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_avs.c
@@ -379,7 +379,7 @@ static int demux_avs_control(demuxer_t *demuxer, int cmd, void *arg)
         case DEMUXER_CTRL_GET_PERCENT_POS:
         {
             if (!AVS->video_info->num_frames) return DEMUXER_CTRL_DONTKNOW;
-            *((int *)arg) = (int) (AVS->frameno * 100 / AVS->video_info->num_frames);
+            *((int *)arg) = AVS->frameno * 100 / AVS->video_info->num_frames;
             return DEMUXER_CTRL_OK;
         }
     default:
